fix(#1139): load the Video.js surface client-only, off the server's module graph
Importing `@videojs/react` pulls in `@videojs/store`, whose `selector.js` builds
a module-scope singleton holding `new AbortController()`. Workerd forbids
constructing an I/O-bound object at global scope, so the mere import threw
"Disallowed operation called within global scope" once per isolate cold start —
~179x/day in production, error-level, and the largest single contributor to the
error baseline every rate alert has to be threshold-tuned around.

`video-player.tsx` already deferred `createPlayer` out of module scope for the
same reason, but that could never fix this one: the throw happens on IMPORT,
before any of our code runs. So the module has to stay out of the server's
evaluated graph entirely.

Move every `@videojs/*` import into `video-player-surface.tsx` and reach it via
a dynamic import gated on a mounted flag. `lazy()` alone is insufficient —
React invokes the loader during SSR too — so the gate, not the laziness, is what
keeps the server out. The skeleton and image-only branches touch no Video.js
code and still render on the server unchanged.

The player subtree no longer emits SSR markup; it mounts after hydration behind
a poster/skeleton placeholder in the same aspect-ratio container, so nothing
shifts. Little is lost in practice: the module-scope throw already degraded that
subtree on the server.

Verified against the built worker (`bun run build` + `wrangler dev`), which is
the only place this reproduces — `bun dev`'s lazy SSR module runner never
evaluates it. One occurrence before, zero after, across /, /login and /pricing,
with the Video.js code emitted to its own chunk that the server entry never
statically imports.
